back-end/requirements.txt
asgiref==3.7.2
attrs==23.1.0
cffi==1.16.0
cryptography==41.0.4
Django==4.2.5
django-model-utils==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
packaging==23.2
pycparser==2.21
PyJWT==1.7.1
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
pathlib==1.0.1
torch>=2.2.1
torchvision>=0.17.1
Pillow>=9.4.0
segmentation_models_pytorch>=0.3.3
six>=1.16.0